requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.6.0
magic-filter==1.0.9
multidict==6.0.4
openai==0.27.5
python-dotenv==0.21.1
python-telegram-bot==20.2
pytz==2023.3
requests==2.29.0
rfc3986==1.5.0
sniffio==1.3.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.9.2
zipp==3.15.0
